requirements.txt
requests
python-dotenv
urllib3>=2.0.0